+++ src/parser/map_output.h
@@ -22,12 +22,15 @@ std::expected<void, std::string> write_tempo_map(
     const std::string& path, const std::vector<TempoMapEntry>& entries);
 
 // Serialize an undisplaced source-frame phase-reset list to the canonical
-// .resetmap text form: one source-frame integer per line, in input order.
-// The companion to write_frame_map on the phase-reset axis — the
-// frame-domain export warptempo_parser emits and the engine-only synthesis
-// driver consumes (read_reset_map in frame_map.h). The caller supplies the
-// already-resolved active source frames (phase_reset_source_frames drops
-// disabled markers and converts time->source frame via nearbyint), so this
-// writer applies no policy: it does not displace, sort, or dedupe.
+// .resetmap text form: one undisplaced source-frame double per line, in
+// input order, at up to 17 significant digits (round-trips an IEEE double
+// exactly, same serialization as write_frame_map; whole-frame values print
+// with no decimal point). The companion to write_frame_map on the
+// phase-reset axis — the frame-domain export warptempo_parser emits and the
+// engine-only synthesis driver consumes (read_reset_map in frame_map.h). The
+// caller supplies the already-resolved active source-frame positions
+// (phase_reset_source_frames drops disabled markers and converts
+// time->exact double source frame), so this writer applies no policy: it
+// does not displace, sort, or dedupe.
 std::expected<void, std::string> write_reset_map(
-    const std::string& path, const std::vector<int64_t>& source_frames);
+    const std::string& path, const std::vector<double>& source_frames);
